requirements.txt
python-telegram-bot[callback-data,webhooks]>=21.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
